Refactor child management package structure and implement Allergy and Guardian services
