+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/AllergyController.java
@@ -1,4 +1,4 @@
-package com.careconnect.coreapi.childmgmt.controller;
+package com.careconnect.coreapi.childmgmt.internal;
 
 import com.careconnect.coreapi.childmgmt.jpa.Allergy;
 import com.careconnect.coreapi.childmgmt.jpa.ChildAllergy;
