+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/GuardianController.java
@@ -1,4 +1,4 @@
-package com.careconnect.coreapi.childmgmt.controller;
+package com.careconnect.coreapi.childmgmt.internal;
 
 import com.careconnect.coreapi.childmgmt.jpa.Guardian;
 import com.careconnect.coreapi.childmgmt.service.GuardianService;
